+++ sdk/resourcemanager/automation/armautomation/python3package_client.go
@@ -23,20 +23,20 @@ import (
 	"strings"
 )
 
-// CertificateClient contains the methods for the Certificate group.
-// Don't use this type directly, use NewCertificateClient() instead.
-type CertificateClient struct {
+// Python3PackageClient contains the methods for the Python3Package group.
+// Don't use this type directly, use NewPython3PackageClient() instead.
+type Python3PackageClient struct {
 	host           string
 	subscriptionID string
 	pl             runtime.Pipeline
 }
 
-// NewCertificateClient creates a new instance of CertificateClient with the specified values.
+// NewPython3PackageClient creates a new instance of Python3PackageClient with the specified values.
 // subscriptionID - Gets subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID
 // forms part of the URI for every service call.
 // credential - used to authorize requests. Usually a credential from azidentity.
 // options - pass nil to accept the default values.
-func NewCertificateClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*CertificateClient, error) {
+func NewPython3PackageClient(subscriptionID string, credential azcore.TokenCredential, options *arm.ClientOptions) (*Python3PackageClient, error) {
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
@@ -48,7 +48,7 @@ func NewCertificateClient(subscriptionID string, credential azcore.TokenCredenti
 	if err != nil {
 		return nil, err
 	}
-	client := &CertificateClient{
+	client := &Python3PackageClient{
 		subscriptionID: subscriptionID,
 		host:           ep,
 		pl:             pl,
@@ -56,33 +56,33 @@ func NewCertificateClient(subscriptionID string, credential azcore.TokenCredenti
 	return client, nil
 }
 
-// CreateOrUpdate - Create a certificate.
+// CreateOrUpdate - Create or Update the python 3 package identified by package name.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-08-08
 // resourceGroupName - Name of an Azure Resource group.
 // automationAccountName - The name of the automation account.
-// certificateName - The parameters supplied to the create or update certificate operation.
-// parameters - The parameters supplied to the create or update certificate operation.
-// options - CertificateClientCreateOrUpdateOptions contains the optional parameters for the CertificateClient.CreateOrUpdate
+// packageName - The name of python package.
+// parameters - The create or update parameters for python package.
+// options - Python3PackageClientCreateOrUpdateOptions contains the optional parameters for the Python3PackageClient.CreateOrUpdate
 // method.
-func (client *CertificateClient) CreateOrUpdate(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, parameters CertificateCreateOrUpdateParameters, options *CertificateClientCreateOrUpdateOptions) (CertificateClientCreateOrUpdateResponse, error) {
-	req, err := client.createOrUpdateCreateRequest(ctx, resourceGroupName, automationAccountName, certificateName, parameters, options)
+func (client *Python3PackageClient) CreateOrUpdate(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, parameters PythonPackageCreateParameters, options *Python3PackageClientCreateOrUpdateOptions) (Python3PackageClientCreateOrUpdateResponse, error) {
+	req, err := client.createOrUpdateCreateRequest(ctx, resourceGroupName, automationAccountName, packageName, parameters, options)
 	if err != nil {
-		return CertificateClientCreateOrUpdateResponse{}, err
+		return Python3PackageClientCreateOrUpdateResponse{}, err
 	}
 	resp, err := client.pl.Do(req)
 	if err != nil {
-		return CertificateClientCreateOrUpdateResponse{}, err
+		return Python3PackageClientCreateOrUpdateResponse{}, err
 	}
 	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusCreated) {
-		return CertificateClientCreateOrUpdateResponse{}, runtime.NewResponseError(resp)
+		return Python3PackageClientCreateOrUpdateResponse{}, runtime.NewResponseError(resp)
 	}
 	return client.createOrUpdateHandleResponse(resp)
 }
 
 // createOrUpdateCreateRequest creates the CreateOrUpdate request.
-func (client *CertificateClient) createOrUpdateCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, parameters CertificateCreateOrUpdateParameters, options *CertificateClientCreateOrUpdateOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/certificates/{certificateName}"
+func (client *Python3PackageClient) createOrUpdateCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, parameters PythonPackageCreateParameters, options *Python3PackageClientCreateOrUpdateOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/python3Packages/{packageName}"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -91,10 +91,10 @@ func (client *CertificateClient) createOrUpdateCreateRequest(ctx context.Context
 		return nil, errors.New("parameter automationAccountName cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{automationAccountName}", url.PathEscape(automationAccountName))
-	if certificateName == "" {
-		return nil, errors.New("parameter certificateName cannot be empty")
+	if packageName == "" {
+		return nil, errors.New("parameter packageName cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{certificateName}", url.PathEscape(certificateName))
+	urlPath = strings.ReplaceAll(urlPath, "{packageName}", url.PathEscape(packageName))
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
@@ -111,39 +111,39 @@ func (client *CertificateClient) createOrUpdateCreateRequest(ctx context.Context
 }
 
 // createOrUpdateHandleResponse handles the CreateOrUpdate response.
-func (client *CertificateClient) createOrUpdateHandleResponse(resp *http.Response) (CertificateClientCreateOrUpdateResponse, error) {
-	result := CertificateClientCreateOrUpdateResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.Certificate); err != nil {
-		return CertificateClientCreateOrUpdateResponse{}, err
+func (client *Python3PackageClient) createOrUpdateHandleResponse(resp *http.Response) (Python3PackageClientCreateOrUpdateResponse, error) {
+	result := Python3PackageClientCreateOrUpdateResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Module); err != nil {
+		return Python3PackageClientCreateOrUpdateResponse{}, err
 	}
 	return result, nil
 }
 
-// Delete - Delete the certificate.
+// Delete - Delete the python 3 package by name.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-08-08
 // resourceGroupName - Name of an Azure Resource group.
 // automationAccountName - The name of the automation account.
-// certificateName - The name of certificate.
-// options - CertificateClientDeleteOptions contains the optional parameters for the CertificateClient.Delete method.
-func (client *CertificateClient) Delete(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, options *CertificateClientDeleteOptions) (CertificateClientDeleteResponse, error) {
-	req, err := client.deleteCreateRequest(ctx, resourceGroupName, automationAccountName, certificateName, options)
+// packageName - The python package name.
+// options - Python3PackageClientDeleteOptions contains the optional parameters for the Python3PackageClient.Delete method.
+func (client *Python3PackageClient) Delete(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, options *Python3PackageClientDeleteOptions) (Python3PackageClientDeleteResponse, error) {
+	req, err := client.deleteCreateRequest(ctx, resourceGroupName, automationAccountName, packageName, options)
 	if err != nil {
-		return CertificateClientDeleteResponse{}, err
+		return Python3PackageClientDeleteResponse{}, err
 	}
 	resp, err := client.pl.Do(req)
 	if err != nil {
-		return CertificateClientDeleteResponse{}, err
+		return Python3PackageClientDeleteResponse{}, err
 	}
-	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return CertificateClientDeleteResponse{}, runtime.NewResponseError(resp)
+	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusNoContent) {
+		return Python3PackageClientDeleteResponse{}, runtime.NewResponseError(resp)
 	}
-	return CertificateClientDeleteResponse{}, nil
+	return Python3PackageClientDeleteResponse{}, nil
 }
 
 // deleteCreateRequest creates the Delete request.
-func (client *CertificateClient) deleteCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, options *CertificateClientDeleteOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/certificates/{certificateName}"
+func (client *Python3PackageClient) deleteCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, options *Python3PackageClientDeleteOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/python3Packages/{packageName}"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -152,10 +152,10 @@ func (client *CertificateClient) deleteCreateRequest(ctx context.Context, resour
 		return nil, errors.New("parameter automationAccountName cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{automationAccountName}", url.PathEscape(automationAccountName))
-	if certificateName == "" {
-		return nil, errors.New("parameter certificateName cannot be empty")
+	if packageName == "" {
+		return nil, errors.New("parameter packageName cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{certificateName}", url.PathEscape(certificateName))
+	urlPath = strings.ReplaceAll(urlPath, "{packageName}", url.PathEscape(packageName))
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
@@ -171,31 +171,31 @@ func (client *CertificateClient) deleteCreateRequest(ctx context.Context, resour
 	return req, nil
 }
 
-// Get - Retrieve the certificate identified by certificate name.
+// Get - Retrieve the python 3 package identified by package name.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-08-08
 // resourceGroupName - Name of an Azure Resource group.
 // automationAccountName - The name of the automation account.
-// certificateName - The name of certificate.
-// options - CertificateClientGetOptions contains the optional parameters for the CertificateClient.Get method.
-func (client *CertificateClient) Get(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, options *CertificateClientGetOptions) (CertificateClientGetResponse, error) {
-	req, err := client.getCreateRequest(ctx, resourceGroupName, automationAccountName, certificateName, options)
+// packageName - The python package name.
+// options - Python3PackageClientGetOptions contains the optional parameters for the Python3PackageClient.Get method.
+func (client *Python3PackageClient) Get(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, options *Python3PackageClientGetOptions) (Python3PackageClientGetResponse, error) {
+	req, err := client.getCreateRequest(ctx, resourceGroupName, automationAccountName, packageName, options)
 	if err != nil {
-		return CertificateClientGetResponse{}, err
+		return Python3PackageClientGetResponse{}, err
 	}
 	resp, err := client.pl.Do(req)
 	if err != nil {
-		return CertificateClientGetResponse{}, err
+		return Python3PackageClientGetResponse{}, err
 	}
 	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return CertificateClientGetResponse{}, runtime.NewResponseError(resp)
+		return Python3PackageClientGetResponse{}, runtime.NewResponseError(resp)
 	}
 	return client.getHandleResponse(resp)
 }
 
 // getCreateRequest creates the Get request.
-func (client *CertificateClient) getCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, options *CertificateClientGetOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/certificates/{certificateName}"
+func (client *Python3PackageClient) getCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, options *Python3PackageClientGetOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/python3Packages/{packageName}"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -204,10 +204,10 @@ func (client *CertificateClient) getCreateRequest(ctx context.Context, resourceG
 		return nil, errors.New("parameter automationAccountName cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{automationAccountName}", url.PathEscape(automationAccountName))
-	if certificateName == "" {
-		return nil, errors.New("parameter certificateName cannot be empty")
+	if packageName == "" {
+		return nil, errors.New("parameter packageName cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{certificateName}", url.PathEscape(certificateName))
+	urlPath = strings.ReplaceAll(urlPath, "{packageName}", url.PathEscape(packageName))
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
@@ -224,27 +224,27 @@ func (client *CertificateClient) getCreateRequest(ctx context.Context, resourceG
 }
 
 // getHandleResponse handles the Get response.
-func (client *CertificateClient) getHandleResponse(resp *http.Response) (CertificateClientGetResponse, error) {
-	result := CertificateClientGetResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.Certificate); err != nil {
-		return CertificateClientGetResponse{}, err
+func (client *Python3PackageClient) getHandleResponse(resp *http.Response) (Python3PackageClientGetResponse, error) {
+	result := Python3PackageClientGetResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Module); err != nil {
+		return Python3PackageClientGetResponse{}, err
 	}
 	return result, nil
 }
 
-// NewListByAutomationAccountPager - Retrieve a list of certificates.
+// NewListByAutomationAccountPager - Retrieve a list of python 3 packages.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-08-08
 // resourceGroupName - Name of an Azure Resource group.
 // automationAccountName - The name of the automation account.
-// options - CertificateClientListByAutomationAccountOptions contains the optional parameters for the CertificateClient.ListByAutomationAccount
+// options - Python3PackageClientListByAutomationAccountOptions contains the optional parameters for the Python3PackageClient.ListByAutomationAccount
 // method.
-func (client *CertificateClient) NewListByAutomationAccountPager(resourceGroupName string, automationAccountName string, options *CertificateClientListByAutomationAccountOptions) *runtime.Pager[CertificateClientListByAutomationAccountResponse] {
-	return runtime.NewPager(runtime.PagingHandler[CertificateClientListByAutomationAccountResponse]{
-		More: func(page CertificateClientListByAutomationAccountResponse) bool {
+func (client *Python3PackageClient) NewListByAutomationAccountPager(resourceGroupName string, automationAccountName string, options *Python3PackageClientListByAutomationAccountOptions) *runtime.Pager[Python3PackageClientListByAutomationAccountResponse] {
+	return runtime.NewPager(runtime.PagingHandler[Python3PackageClientListByAutomationAccountResponse]{
+		More: func(page Python3PackageClientListByAutomationAccountResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
-		Fetcher: func(ctx context.Context, page *CertificateClientListByAutomationAccountResponse) (CertificateClientListByAutomationAccountResponse, error) {
+		Fetcher: func(ctx context.Context, page *Python3PackageClientListByAutomationAccountResponse) (Python3PackageClientListByAutomationAccountResponse, error) {
 			var req *policy.Request
 			var err error
 			if page == nil {
@@ -253,14 +253,14 @@ func (client *CertificateClient) NewListByAutomationAccountPager(resourceGroupNa
 				req, err = runtime.NewRequest(ctx, http.MethodGet, *page.NextLink)
 			}
 			if err != nil {
-				return CertificateClientListByAutomationAccountResponse{}, err
+				return Python3PackageClientListByAutomationAccountResponse{}, err
 			}
 			resp, err := client.pl.Do(req)
 			if err != nil {
-				return CertificateClientListByAutomationAccountResponse{}, err
+				return Python3PackageClientListByAutomationAccountResponse{}, err
 			}
 			if !runtime.HasStatusCode(resp, http.StatusOK) {
-				return CertificateClientListByAutomationAccountResponse{}, runtime.NewResponseError(resp)
+				return Python3PackageClientListByAutomationAccountResponse{}, runtime.NewResponseError(resp)
 			}
 			return client.listByAutomationAccountHandleResponse(resp)
 		},
@@ -268,8 +268,8 @@ func (client *CertificateClient) NewListByAutomationAccountPager(resourceGroupNa
 }
 
 // listByAutomationAccountCreateRequest creates the ListByAutomationAccount request.
-func (client *CertificateClient) listByAutomationAccountCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, options *CertificateClientListByAutomationAccountOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/certificates"
+func (client *Python3PackageClient) listByAutomationAccountCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, options *Python3PackageClientListByAutomationAccountOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/python3Packages"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -294,40 +294,40 @@ func (client *CertificateClient) listByAutomationAccountCreateRequest(ctx contex
 }
 
 // listByAutomationAccountHandleResponse handles the ListByAutomationAccount response.
-func (client *CertificateClient) listByAutomationAccountHandleResponse(resp *http.Response) (CertificateClientListByAutomationAccountResponse, error) {
-	result := CertificateClientListByAutomationAccountResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.CertificateListResult); err != nil {
-		return CertificateClientListByAutomationAccountResponse{}, err
+func (client *Python3PackageClient) listByAutomationAccountHandleResponse(resp *http.Response) (Python3PackageClientListByAutomationAccountResponse, error) {
+	result := Python3PackageClientListByAutomationAccountResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.ModuleListResult); err != nil {
+		return Python3PackageClientListByAutomationAccountResponse{}, err
 	}
 	return result, nil
 }
 
-// Update - Update a certificate.
+// Update - Update the python 3 package identified by package name.
 // If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-08-08
 // resourceGroupName - Name of an Azure Resource group.
 // automationAccountName - The name of the automation account.
-// certificateName - The parameters supplied to the update certificate operation.
-// parameters - The parameters supplied to the update certificate operation.
-// options - CertificateClientUpdateOptions contains the optional parameters for the CertificateClient.Update method.
-func (client *CertificateClient) Update(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, parameters CertificateUpdateParameters, options *CertificateClientUpdateOptions) (CertificateClientUpdateResponse, error) {
-	req, err := client.updateCreateRequest(ctx, resourceGroupName, automationAccountName, certificateName, parameters, options)
+// packageName - The name of python package.
+// parameters - The update parameters for python package.
+// options - Python3PackageClientUpdateOptions contains the optional parameters for the Python3PackageClient.Update method.
+func (client *Python3PackageClient) Update(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, parameters PythonPackageUpdateParameters, options *Python3PackageClientUpdateOptions) (Python3PackageClientUpdateResponse, error) {
+	req, err := client.updateCreateRequest(ctx, resourceGroupName, automationAccountName, packageName, parameters, options)
 	if err != nil {
-		return CertificateClientUpdateResponse{}, err
+		return Python3PackageClientUpdateResponse{}, err
 	}
 	resp, err := client.pl.Do(req)
 	if err != nil {
-		return CertificateClientUpdateResponse{}, err
+		return Python3PackageClientUpdateResponse{}, err
 	}
 	if !runtime.HasStatusCode(resp, http.StatusOK) {
-		return CertificateClientUpdateResponse{}, runtime.NewResponseError(resp)
+		return Python3PackageClientUpdateResponse{}, runtime.NewResponseError(resp)
 	}
 	return client.updateHandleResponse(resp)
 }
 
 // updateCreateRequest creates the Update request.
-func (client *CertificateClient) updateCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, certificateName string, parameters CertificateUpdateParameters, options *CertificateClientUpdateOptions) (*policy.Request, error) {
-	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/certificates/{certificateName}"
+func (client *Python3PackageClient) updateCreateRequest(ctx context.Context, resourceGroupName string, automationAccountName string, packageName string, parameters PythonPackageUpdateParameters, options *Python3PackageClientUpdateOptions) (*policy.Request, error) {
+	urlPath := "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Automation/automationAccounts/{automationAccountName}/python3Packages/{packageName}"
 	if resourceGroupName == "" {
 		return nil, errors.New("parameter resourceGroupName cannot be empty")
 	}
@@ -336,10 +336,10 @@ func (client *CertificateClient) updateCreateRequest(ctx context.Context, resour
 		return nil, errors.New("parameter automationAccountName cannot be empty")
 	}
 	urlPath = strings.ReplaceAll(urlPath, "{automationAccountName}", url.PathEscape(automationAccountName))
-	if certificateName == "" {
-		return nil, errors.New("parameter certificateName cannot be empty")
+	if packageName == "" {
+		return nil, errors.New("parameter packageName cannot be empty")
 	}
-	urlPath = strings.ReplaceAll(urlPath, "{certificateName}", url.PathEscape(certificateName))
+	urlPath = strings.ReplaceAll(urlPath, "{packageName}", url.PathEscape(packageName))
 	if client.subscriptionID == "" {
 		return nil, errors.New("parameter client.subscriptionID cannot be empty")
 	}
@@ -356,10 +356,10 @@ func (client *CertificateClient) updateCreateRequest(ctx context.Context, resour
 }
 
 // updateHandleResponse handles the Update response.
-func (client *CertificateClient) updateHandleResponse(resp *http.Response) (CertificateClientUpdateResponse, error) {
-	result := CertificateClientUpdateResponse{}
-	if err := runtime.UnmarshalAsJSON(resp, &result.Certificate); err != nil {
-		return CertificateClientUpdateResponse{}, err
+func (client *Python3PackageClient) updateHandleResponse(resp *http.Response) (Python3PackageClientUpdateResponse, error) {
+	result := Python3PackageClientUpdateResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Module); err != nil {
+		return Python3PackageClientUpdateResponse{}, err
 	}
 	return result, nil
 }